+++ CHANGELOG.md
@@ -10,6 +10,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - parallel tuning on single node using Noodles runner
 - option to pass new defaults for block dimensions
+- option to pass a Python function as code generator
 
 ## [0.1.4] - 2017-06-14
 ### Changed